fix(#729): move production-story unit lib+test out of [id] route dir
CI's changed-spec runner does `jest --runTestsByPath '<path>'`. When a unit
spec lives under a Next-style `[id]` dynamic-route dir, jest interprets the
`[id]` as a regex char class → 0 matches → exit 1, so the spec (and the
products.ts fetch change it guards) never actually ran — a false-negative CI
failure.

Relocate the pure lib + its unit test to a NON-bracketed sibling dir
(production-story-lib/) and update the route import. Route stays under [id]
(it's the dynamic route). No story logic changed — pure file move.

GOTCHA: unit tests must never live under a `[param]` route dir.
